+++ expyriment/__init__.py
@@ -36,7 +36,7 @@ __date__ = ''
 
 
 import sys as _sys
-from ._internals import get_version
+from ._internals import get_version, import_all_extras
 from ._internals import PYTHON3 as _PYTHON3
 
 if not( (_sys.version_info[0] == 2 and _sys.version_info[1] >= 6) or
@@ -117,5 +117,10 @@ stimuli.Video.get_ffmpeg_binary()  # in case it needs to be downloaded
 if not misc.is_android_running():
     from ._api_reference_tool import show_documentation
 
+# add extras folder if it exists
+if _internals.get_plugins_folder() is not None:
+    _sys.path.append(_internals.get_plugins_folder())
+
+# post import hook
 exec(_internals.post_import_hook())