+++ expyriment/design/permute/__init__.py
@@ -1,0 +1,16 @@
+"""
+The permute module.
+
+This module implements permutation of blocks, trials and conditions.
+
+"""
+
+__author__ = 'Florian Krause <florian@expyriment.org>, \
+Oliver Lindemann <oliver@expyriment.org>'
+__version__ = ''
+__revision__ = ''
+__date__ = ''
+
+
+from ._permute import is_permutation_type
+from ._permute import balanced_latin_square, cycled_latin_square