+++ expyriment/io/extras/_midiout_defaults.py
@@ -1,3 +1,0 @@
-# MidiOut defaults
-midiout_buffer_size = 1024
-midiout_latency = 0