+++ expyriment/io/extras/_tbvnetworkinterface_defaults.py
@@ -1,3 +1,0 @@
-# TbvNetworkInterface defaults
-tbvnetworkinterface_timeout = 2000
-tbvnetworkinterface_connect = True