+++ expyriment/io/extras/_tcpclient_defaults.py
@@ -1,3 +1,0 @@
-# TcpClient defaults
-tcpclient_default_package_size = 1024
-tcpclient_connect = True