+++ expyriment/misc/constants.py
@@ -3,8 +3,6 @@
 This module contains expyriment constants.
 
 """
-from __future__ import absolute_import, print_function, division
-from builtins import *
 
 __author__ = 'Florian Krause <florian@expyriment.org>, \
 Oliver Lindemann <oliver@expyriment.org>'