+++ expyriment/misc/extras/__init__.py
@@ -15,19 +15,18 @@ __revision__ = ''
 __date__ = ''
 
 
-import os as _os
-
-from . import defaults
 from ... import _internals
 
+print("Misc plugins:")
+for name, code in _internals.import_plugins_code("misc").items():
+    print(" " + name)
+    try:
+        exec(code)
+    except Exception as err:
+        print("Warning: Could not import {0}".format(name))
+        print(" {0}".format(err))
 
-for _plugins in [_internals.import_plugins(__file__),
-                _internals.import_plugins_from_settings_folder(__file__)]:
-    for _plugin in _plugins:
-        try:
-            exec(_plugins[_plugin])
-        except Exception as err:
-            print("Warning: Could not import {0}".format(
-                _os.path.dirname(__file__) + _os.sep + _plugin))
-            print(" {0}".format(err))
-
+try:
+    del (name, code)
+except:
+    pass