+++ expyriment/stimuli/extras/_noisetone_defaults.py
@@ -1,4 +1,0 @@
-# NoiseTone
-noisetone_samplerate = 44100
-noisetone_bitdepth = 16
-noisetone_amplitude = 0.5