+++ expyriment/stimuli/extras/_polygondot_defaults.py
@@ -1,5 +1,0 @@
-# PolygonDot
-polygondot_colour = None # 'None' is experiment_text_colour
-polygondot_position = (0, 0)
-polygondot_anti_aliasing = 10
-polygondot_resolution_factor = 1