+++ expyriment/stimuli/extras/_polygonellipse_defaults.py
@@ -1,6 +1,0 @@
-# Polygon Ellipse
-polygonellipse_colour = None # 'None' is experiment_text_colour
-polygonellipse_line_width = 0
-polygonellipse_position = (0, 0)
-polygonellipse_anti_aliasing = 5
-polygonellipse_resolution_factor = 1