+++ expyriment/stimuli/extras/_polygonrectangle_defaults.py
@@ -1,3 +1,0 @@
-polygonrectangle_colour = None # 'None' is experiment_text_colour
-polygonrectangle_position = (0, 0)
-polygonrectangle_anti_aliasing = 0