+++ expyriment/stimuli/extras/_randomdotkinematogram_defaults.py
@@ -1,7 +1,0 @@
-randomdotkinematogram_dot_speed = 100
-randomdotkinematogram_dot_lifetime = 400
-randomdotkinematogram_dot_radius = 3
-randomdotkinematogram_dot_colour = None # expyriment default
-randomdotkinematogram_background_colour = None # expyriment default
-randomdotkinematogram_north_up_clockwise = True
-randomdotkinematogram_position = (0,0)